(SERVER-1501) Refactor graph class to use interface
The current Java class that we have to represent graph data assumes
that you will always be pulling the data out of a gatling simulation.
We now have some use cases where we want to graph data that isn't
strictly related to gatling, and covers the whole jenkins job run
rather than just an individual gatling sim.

This commit simply refactors the Graph class to provide an interface
that we can use more generally, and the concrete class is now named
`SimulationGraph`, to make it more explicit that it is coupled to
a gatling simulation.
